+++ cvxpy/constraints/leq_constraint.py
@@ -20,7 +20,8 @@ along with CVXPY.  If not, see <http://www.gnu.org/licenses/>.
 import cvxpy.utilities as u
 import cvxpy.lin_ops.lin_utils as lu
 # Only need Variable from expressions, but that would create a circular import.
-from cvxpy import expressions
+from cvxpy.expressions import cvxtypes
+import abc
 from cvxpy.constraints.constraint import Constraint
 import numpy as np
 
@@ -30,7 +31,7 @@ class LeqConstraint(u.Canonical, Constraint):
     def __init__(self, lh_exp, rh_exp):
         self.args = [lh_exp, rh_exp]
         self._expr = lh_exp - rh_exp
-        self.dual_variable = expressions.variables.Variable(*self._expr.size)
+        self.dual_variable = cvxtypes.variable()(*self._expr.size)
         super(LeqConstraint, self).__init__()
 
     @property
@@ -120,10 +121,21 @@ class LeqConstraint(u.Canonical, Constraint):
         -------
         bool
         """
-        if self._expr.value is None:
+        resid = self.residual.value
+        if resid is None:
             return None
         else:
-            return np.all(self._expr.value <= self.TOLERANCE)
+            return np.all(resid <= self.TOLERANCE)
+
+    @property
+    def residual(self):
+       """The residual of the constraint.
+
+       Returns
+       -------
+       Expression
+       """
+       return cvxtypes.pos()(self._expr)
 
     @property
     def violation(self):
@@ -133,10 +145,7 @@ class LeqConstraint(u.Canonical, Constraint):
         -------
         NumPy matrix
         """
-        if self._expr.value is None:
-            return None
-        else:
-            return np.maximum(self._expr.value, 0)
+        return self.residual.value
 
     # The value of the dual variable.
     @property